requirements.txt
asgiref==3.7.2
astroid==2.4.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
Faker==19.6.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.5.5
typing_extensions==4.8.0
validate-docbr==1.10.0
wrapt==1.12.1
